Introducing the C Parser
Well, we've put the C parser in and made it the default.  It is
mostly functionally complete but it may break a few things here
and there.  It certainly needs a bit of cleanup as well.  We'll
work on this over the course of time.  The current protocol is
renamed Kharon::Protocol::ArrayHashPerl so that we can still use
it in tests...