Fix timeouts on connect (hopefully, subject to fuller testing).
In Kharon::Engine::Client::NetImpl, we were assuming that the
IO::Socket timeout would affect connect, read, and write.  In fact,
it affects connect.  However, it affects neither read nor write.
So, we implement a select-based read timeout and neglect the write
for now.  We should get back to that later.

We also reduce the timeout values to be quite a bit shorter by
default.  Perhaps not short enough for the modern world, however.