Make scalar methods that ``return'' return undef or ().
When evaluating a function defined with mk_scalar_method() we convert
an empty list into undef in a scalar context and () in a list context.
We do this because in perl, return with no arguments will return () in
a list context which is how we evaluate things on the server side.  We
used to toss an exception if the returned list had length != 1 but this
is suboptimal as it is valid for functions to ``return;'' which should
be converted to undef.  So, now we support length == 0 and munge the
output a bit to make it make some level of sense.

We also added unit tests to ensure that the functionality works.