Provide better mechanisms for defining available methods.
If an object that we are presenting defines:

	@KHARON_RO_SC_EXPORT
	@KHARON_RO_AC_EXPORT
	@KHARON_RW_SC_EXPORT
	@KHARON_RW_AC_EXPORT

We use these to decide what methods are available and which ones
will issue referrals rather than requiring the daemon to pass them
to use during our initialisation.  We also look for a class method
&KHARON_MASTER() which we will call expecting it to return the
hostname of the master.  We will refer the RW commands above to
the result of &KHARON_MASTER if it is different than hostname().

We preserve the ability to send cmds, refercmds, next_server and
master as arguments when constructing the server object for backwards
compatibility.