We extend Kharon::dbutils a bit.
We needed to be able to deal with lists where the list key has the
same name as the primary key of the original table.  In the prior
code, this caused some level of issue, so we ran around prepending
<table>. to all manner of things.

We also need to be able to rename the target of lists so we added
a paramater to do that.  We may very well expand this to be able to
be used for scalar fields but it's less compelling at this point.