We modify the semantics of our Entitlement:: objects.
We modify the semantics of the check method to take a verb and
arguments rather than a list of ``entitlements''.  We can think of
the verb as the prior entitlement argument but the arguments allow
us to construct entitlement objects that modify their behaviour
based on the object of the call rather than just the verb.

Now, derived classes should define a check1() method which accepts
not only the ``entitlement'' but rather the ``verb'' and ``arguments''.
The return value of check1() is extended as well.  As well as the
prior 0 meaning ``Permission denied'' and 1 meaning allowed, we also
allow check1() to return a string which will be used to form the error
message if we are throwing an exception.  If we aren't tossing exceptions
then the string will be converted into a false return code.

We also can optionally throw an exception when permission is denied.
This is more in keeping with the way that krb5_admin, et al. work.