Implement Kharon::Entitlement::Object.
Implement a ACL that takes an object and will invoke methods in it
to return the entitlement status.  Kharon::Entitlement::Object when
called will look for methods of the form KHARON_COMMON_ACL and
KHARON_ACL_<verb>.  It will first invoke KHARON_COMMON_ACL giving
it the arguments ($verb, @args).  If this returns a defined value,
it will be used as the entitlement status.  If not, then it will
invoke KHARON_ACL_<verb> in the same way.  If this is not found,
then it will return 0 (Permission denied).