Add the ability to add an acl (Kharon::Entitlement) to a server.
This will be evaluated prior to calling the function.  The server
will pass along the command name and all of its arguments for the
entitlement code to evaluate.  We export a new method set_acl() to
allow the acl to be set after server creation.  And we move the
evaluation of exitcmds to do_command as we need to run them without
requiring the user be ACLed---after all, you shouldn't need an
entitlement to shutdown the connexion.