Fix 66e5d34fe6ded415c3ee16fbf65b459e14314d2b differently.
In 66e5d34fe6ded415c3ee16fbf65b459e14314d2b, we evaluate a loop in
Kharon::Entitlement::Equals over the supplied $verb and @predicate
to fix an issue with how we are called from Kharon::Entitlement::ACLFile.
But the bug is in the latter which was not updated to use the new
semantics that ->check() takes $verb, @predicate rather than a list
of ``entities''.  We move the fix into ACLFile.