Change Kharon::InputValidation to return array refs and undef.
On Viktor Dukhovni's suggestion, having the input validation functions
return the new list and assuming an empty list means a lack of modification
is a bit broken.  We now require the validate method to return either undef
meaning no mods or an array ref giving the entirety of the new arguments.

We also update a few comments to be a bit better.